refactor: extract shared companion-chat pipeline from both pages
The app and overlay pages each carried ~180 near-identical lines of the
send/stream/process pipeline, and they had already drifted — the overlay
forgot to filter empty messages before sending. All of it now lives in
one sendCompanionMessage() module: prompt building, both transports
(direct provider call vs. server route), post-turn processing, keepsakes,
TTS, and the talking animation. Each page's handleSend is now a thin
call passing hooks for its own reactive state.

Fixes carried in the consolidation:
- empty-message filter now applied consistently (was missing in overlay)
- the server-route SSE reader lock is always released, even on an e:
  error line (previously leaked)
- relationship-time state (streak/daysKnown) only updates once the
  character has loaded, closing a race where an early message mutated
  the default state that load then discarded

Net ~330 lines of duplication removed; dead imports pruned from both pages.
