+++ src/routes/app/+page.svelte
@@ -11,30 +11,23 @@
 	import MemoryGraphModal from '$lib/components/memory/MemoryGraphModal.svelte';
 	import { vrmStore } from '$lib/stores/vrm.svelte';
 	import { chatStore } from '$lib/stores/chat.svelte';
-	import { settingsStore } from '$lib/stores/settings.svelte';
 	import { modulesStore } from '$lib/stores/modules.svelte';
-	import { ttsStore } from '$lib/stores/tts.svelte';
 	import { characterStore } from '$lib/stores/character.svelte';
 	import { personaStore } from '$lib/stores/persona.svelte';
 	import { debugEventsStore } from '$lib/stores/debugEvents.svelte';
-	import { getLLMProvider, getTTSProvider, providerSupportsVision } from '$lib/services/providers/registry';
+	import { getLLMProvider, providerSupportsVision } from '$lib/services/providers/registry';
 	import { isLocalLLMProvider } from '$lib/services/providers/local-endpoints';
 	import { canShowImages } from '$lib/services/providers/vision';
-	import { streamChatDirect } from '$lib/services/chat/client-chat';
-	import { processCompanionTurn } from '$lib/services/chat/companion-turn';
-	import { keepImage, type PreparedImage } from '$lib/services/storage/keepsakes';
-	import { type ContentPart, toOpenAIContent } from '$lib/services/chat/content';
+	import { sendCompanionMessage } from '$lib/services/chat/companion-chat';
+	import { type PreparedImage } from '$lib/services/storage/keepsakes';
 	import { isTauri } from '$lib/services/platform';
 	import { browser } from '$app/environment';
-	import type { TTSProvider } from '$lib/types';
 	import type { StateUpdates } from '$lib/types/character';
 	import type { EventDefinition } from '$lib/types/events';
 	import { pop, fadeFast } from '$lib/utils/motion';
 
 	// V2 companion system imports
-	import { buildSystemPrompt, type PromptContext } from '$lib/ai/prompt-builder';
 	import {
-		retrieveRelevantContext,
 		hydrateWorkingMemory,
 		backfillEmbeddings,
 		getEmbeddingBackfillStatus
@@ -129,220 +122,15 @@
 	});
 
 
-	// Build system prompt
-	async function buildCompanionSystemPrompt(userMessage: string, hasImages = false): Promise<string> {
-		const state = characterStore.state;
-		const persona = personaStore.activeCard;
-
-		const memories = await retrieveRelevantContext(userMessage);
-
-		if (import.meta.env.DEV) {
-			console.log('%c[memory hydration]', 'color:#a855f7;font-weight:bold', {
-				relevantFacts: memories.relevantFacts.map((f) => f.content),
-				triggered: memories.triggeredMemories.map((m) => m.content),
-				recentTurns: memories.recentTurns.length
-			});
-		}
-
-		const context: PromptContext = {
-			persona,
-			state,
-			memories,
-			userMessage,
-			systemTime: new Date(),
-			hasImages
-		};
-
-		return buildSystemPrompt(context);
-	}
-
-	// Handle send message
+	// Send a message through the shared companion pipeline.
 	async function handleSend(content: string, images: PreparedImage[] = []) {
-		if ((!content.trim() && images.length === 0) || chatStore.isLoading) return;
-
-		// Check if chat is enabled
-		if (!modulesStore.isModuleEnabled('consciousness')) {
-			chatStore.setError('Chat is disabled. Enable it in Settings > Character > AI Services.');
-			return;
-		}
-
-		const shown = images.map((img) => ({ id: img.id, url: URL.createObjectURL(img.blob) }));
-		chatStore.addMessage('user', content, shown);
-		thinkingImages = shown;
-		chatStore.setLoading(true);
-		chatStore.setError(null);
-		isTyping = true;
-		latestResponse = '';
-
-		characterStore.updateStreak();
-		characterStore.updateDaysKnown();
-
-		try {
-			const consciousnessSettings = modulesStore.getModuleSettings('consciousness');
-			const provider = consciousnessSettings.activeProvider as string;
-			const model = consciousnessSettings.activeModel as string;
-
-			if (!provider) {
-				throw new Error('Please configure a provider in Settings > Modules > Consciousness');
-			}
-
-			const systemPrompt = await buildCompanionSystemPrompt(content, images.length > 0);
-			const providerConfig = settingsStore.getProviderConfig(provider);
-			const apiKey = providerConfig.apiKey;
-			const providerMeta = getLLMProvider(provider);
-
-			if (providerMeta?.requiresApiKey && !apiKey) {
-				throw new Error(`Please configure API key for ${providerMeta.name} in Settings > Providers`);
-			}
-
-			chatStore.addMessage('assistant', '');
-			let fullContent = '';
-			const selectedModel = model || providerMeta?.models?.[0]?.id || '';
-
-			const shouldUseDirectChat = isTauri() || !!providerMeta?.isLocal;
-
-			// The current turn carries the image bytes; prior turns stay text.
-			const history = chatStore.messages.slice(0, -1).filter((m) => m.content || m.images?.length);
-			const directMessages = history.map((m, idx) => {
-				const isCurrentTurn = idx === history.length - 1 && images.length > 0;
-				if (!isCurrentTurn) {
-					return { role: m.role as 'user' | 'assistant', content: m.content };
-				}
-				const parts: ContentPart[] = [];
-				if (m.content) parts.push({ type: 'text', text: m.content });
-				for (const img of images) {
-					parts.push({ type: 'image', mimeType: img.mimeType, data: img.base64 });
-				}
-				return { role: m.role as 'user' | 'assistant', content: parts };
-			});
-
-			if (shouldUseDirectChat) {
-				await new Promise<void>((resolve, reject) => {
-					streamChatDirect(
-						{
-							messages: directMessages,
-							provider: provider as import('$lib/types').LLMProvider,
-							model: selectedModel,
-							apiKey: apiKey || undefined,
-							baseURL: providerConfig.baseUrl || providerMeta?.defaultBaseUrl,
-							systemPrompt
-						},
-						(text) => { fullContent += text; chatStore.updateLastMessage(fullContent); },
-						(error) => reject(new Error(error)),
-						() => resolve()
-					);
-				});
-			} else {
-				const response = await fetch('/api/chat', {
-					method: 'POST',
-					headers: { 'Content-Type': 'application/json' },
-					body: JSON.stringify({
-						messages: directMessages.map((m) => ({
-							role: m.role,
-							content: toOpenAIContent(m.content)
-						})),
-						provider,
-						model: selectedModel,
-						apiKey: apiKey || 'not-needed',
-						baseURL: providerConfig.baseUrl || providerMeta?.defaultBaseUrl,
-						systemPrompt
-					})
-				});
-
-				if (!response.ok) {
-					const errBody = await response.json().catch(() => null);
-					throw new Error(errBody?.error || 'Failed to get response');
-				}
-
-				const reader = response.body?.getReader();
-				const decoder = new TextDecoder();
-				if (!reader) throw new Error('No response body');
-
-				const processLine = (line: string) => {
-					if (line.startsWith('0:')) {
-						const text = JSON.parse(line.slice(2));
-						fullContent += text;
-						chatStore.updateLastMessage(fullContent);
-					} else if (line.startsWith('e:')) {
-						const { error } = JSON.parse(line.slice(2));
-						throw new Error(error);
-					}
-				};
-
-				// Buffer partial lines so a delta split across network chunks doesn't break JSON.parse
-				let streamBuffer = '';
-				while (true) {
-					const { done, value } = await reader.read();
-					if (done) break;
-
-					streamBuffer += decoder.decode(value, { stream: true });
-					const lines = streamBuffer.split('\n');
-					streamBuffer = lines.pop() || '';
-					for (const line of lines) {
-						processLine(line);
-					}
-				}
-				streamBuffer += decoder.decode();
-				if (streamBuffer) processLine(streamBuffer);
-			}
-
-			isTyping = false;
-			const turn = await processCompanionTurn({
-				userMessage: content,
-				companionResponse: fullContent,
-				llm: {
-					provider,
-					model: selectedModel,
-					apiKey: apiKey || undefined,
-					baseURL: providerConfig.baseUrl || providerMeta?.defaultBaseUrl,
-					hasImages: images.length > 0
-				},
-				debug: import.meta.env.DEV
-			});
-			const cleanedResponse = turn.dialogue;
-			lastNewMemory = turn.newMemory;
-			if (turn.triggeredEvent) activeEvent = turn.triggeredEvent;
-
-			// She's seen it and responded; keep the shown images as local keepsakes
-			if (images.length > 0) {
-				await Promise.all(
-					images.map((img) =>
-						keepImage(img.id, img.blob, { mimeType: img.mimeType, note: lastNewMemory })
-					)
-				);
-			}
-
-			chatStore.updateLastMessage(cleanedResponse);
-			latestResponse = cleanedResponse;
-
-			// Trigger talking animation based on response length
-			if (cleanedResponse) {
-				vrmStore.startTalking(cleanedResponse);
-			}
-
-			// TTS - speak if module is enabled
-			const speechState = modulesStore.getModuleState('speech');
-			const speechSettings = modulesStore.getModuleSettings('speech');
-			if (speechState?.enabled && cleanedResponse) {
-				const ttsProvider = speechSettings.activeProvider as TTSProvider;
-				const ttsConfig = settingsStore.getProviderConfig(ttsProvider);
-				const ttsMeta = getTTSProvider(ttsProvider);
-
-				ttsStore.speak(cleanedResponse, {
-					provider: ttsProvider,
-					apiKey: ttsConfig.apiKey,
-					voiceId: speechSettings.activeVoiceId as string || ttsConfig.voiceId,
-					model: speechSettings.activeModel as string || ttsConfig.modelId,
-					baseUrl: ttsConfig.baseUrl || ttsMeta?.defaultBaseUrl,
-					speed: speechSettings.speed as number ?? 1
-				});
-			}
-		} catch (err) {
-			chatStore.setError(err instanceof Error ? err.message : 'Unknown error');
-			isTyping = false;
-		} finally {
-			chatStore.setLoading(false);
-		}
+		await sendCompanionMessage(content, images, {
+			setTyping: (v) => (isTyping = v),
+			setLatestResponse: (v) => (latestResponse = v),
+			setActiveEvent: (e) => (activeEvent = e),
+			onShownImages: (shown) => (thinkingImages = shown),
+			onNewMemory: (m) => (lastNewMemory = m)
+		});
 	}
 
 	// Handle speech bubble hide
